+++ cdl_convert/collection.py
@@ -19,7 +19,7 @@ ColorCorrectionCollection and ColorDecisionList.
 The MIT License (MIT)
 
 cdl_convert
-Copyright (c) 2014 Sean Wallitsch
+Copyright (c) 2015 Sean Wallitsch
 http://github.com/shidarin/cdl_convert/
 
 Permission is hereby granted, free of charge, to any person obtaining a copy