+++ docs/changelog.rst
@@ -2,6 +2,13 @@
 Changelog
 #########
 
+
+Version 0.8
+===========
+
+- Added `--single` flag. When provided with an output collection format, each color correction in the input will be exported to it's own collection.
+- Giving a ColorCorrection a non-duplicate ID now works unless the ``--halt`` flag is given. This means that incoming collections that contain duplicate IDs will not fail out.
+
 Version 0.7.1
 =============